refactor(dashboard): split shared.ts queries into domain-specific modules and update server actions
