+++ src/actions/dashboard/admin.ts
@@ -2,28 +2,32 @@
 
 import { getAuthenticatedUser } from '@/actions/auth';
 import { getWriteOffsLedger } from '@/actions/financials';
+import { assertAdmin } from './queries/auth';
+import { getCachedDashboardKpiMetrics } from './queries/kpis';
 import {
-  assertAdmin,
-  getCachedDashboardKpiMetrics,
   getCachedInventoryStatus,
   getCachedDepartmentAllocation,
   getOverdueReturnsInternal,
   getPendingDisposalsInternal,
   getHighMaintenanceAssetsInternal,
-  getRecentActivitiesInternal,
+} from './queries/inventory';
+import { getRecentActivitiesInternal } from './queries/activities';
+import {
   getDashboardTopHighValueAssetsInternal,
   getDashboardSoftwareOptimizationInternal,
-  type DashboardKpiMetrics,
-  type InventoryStatusResponse,
-  type DepartmentAllocationItem,
-  type OverdueReturnRow,
-  type PendingDisposalRow,
-  type HighMaintenanceRow,
-  type RecentActivity,
-  type TopHighValueAssetRow,
-  type SoftwareOptimizationRow,
-  type WriteOffLedgerRow,
-} from './shared';
+} from './queries/financials';
+import type {
+  DashboardKpiMetrics,
+  InventoryStatusResponse,
+  DepartmentAllocationItem,
+  OverdueReturnRow,
+  PendingDisposalRow,
+  HighMaintenanceRow,
+  RecentActivity,
+  TopHighValueAssetRow,
+  SoftwareOptimizationRow,
+  WriteOffLedgerRow,
+} from '@/types/dashboard';
 
 export interface AdminDashboardBatchData {
   kpiMetrics: DashboardKpiMetrics;
